requirement.txt
h5py
numpy
pyyaml
pprint
enum34
future
scipy
matplotlib
transforms3d
scikit_image
lmdb
setuptools
cffi
easydict
plyfile
pillow
glumpy
pandas
libjasper-dev
